[dns] filter - don't notify on drop if pool is large - uniform _LAST times
If there are more records in our pool (due to round-robin or certain
types of load-balancing DNS) than are live in DNS, then don’t notify if
a record is dropped, because it’s likely to be re-added later on. (It’s
always retained in pool.)

New - if the pool has more records than are live, then send
notification when we drop a record from the pool. If a non-pooling DNS
is being used, then our internal pool size will -equal- the number of
live records, in which case we -do- notify when one drops.

When updating a records _LAST (seen) time, use the same -current- time
for every record we update, and in notifications.