Third big checkin 'Suppress email notifications'
Show a strip chart on the ‘suppress’ page only if user is already (HTTP
Basic) authenticated within the site. That’s because the chart code is
protected by authentication, so don’t want to show a 401 if user isn’t
logged in. Would look very messy. Note problems with
$_SERVER[‘REMOTE_USER’] which often doesn’t work.) Copy d3 CSS
directory to docroot and use it from there. Note also that jQuery was
added to the ‘suppress’ page in a previous code checkin. Display full
‘hours and minutes’ remaining for suppressions - has a nicer feel to
it. Email will contain message if code couldn’t reach the locking
mechanism when checking for suppression.